+++ builder_pkgs/mockito/test/nnbd_support_test.dart
@@ -26,8 +26,8 @@ class Foo {
 class MockFoo extends Mock implements Foo {
   @override
   String returnsNonNullableString() {
-    return super.noSuchMethod(
-        Invocation.method(#returnsNonNullableString, []), 'Dummy') as String;
+    return super.noSuchMethod(Invocation.method(#returnsNonNullableString, []),
+        returnValue: 'Dummy') as String;
   }
 }